Remove PR comment if no changes found
