--- src/upsertPrComment.ts
@@ -1,40 +1,0 @@
-type Octokit = ReturnType<typeof import('@actions/github')['getOctokit']>
-
-export async function upsertPrComment(
-  github: Octokit,
-  repo: {
-    owner: string
-    repo: string
-  },
-  pullRequestNumber: number,
-  title: string,
-  body: string
-): Promise<void> {
-  const existingCommentsResponse = await github.issues.listComments({
-    ...repo,
-    issue_number: pullRequestNumber
-  })
-
-  if (body.length >= 65536) {
-    const truncated = '\n\n\n...long comment body truncated...'
-    body = body.substr(0, 65536 - truncated.length) + truncated
-  }
-
-  const [existingComment] = existingCommentsResponse.data.filter(x =>
-    x.body?.startsWith(title)
-  )
-
-  if (existingComment) {
-    await github.issues.updateComment({
-      ...repo,
-      comment_id: existingComment.id,
-      body
-    })
-  } else {
-    await github.issues.createComment({
-      ...repo,
-      issue_number: pullRequestNumber,
-      body
-    })
-  }
-}
